requirements.txt
pytest
fastapi
pytest-postgresql
requests
sqlalchemy
pydantic
factory-boy
faker
black
isort
